requirements.txt
# Federated AI Sentinel Dependencies
# All versions pinned for reproducibility

# Core ML frameworks
torch==2.1.2
scikit-learn==1.3.2
xgboost==2.0.3
numpy==1.26.3
pandas==2.1.4
polars==0.20.3

# Federated learning
flwr==1.6.0
flwr-datasets==0.0.2

# Deep learning utilities
transformers==4.36.2
pytorch-lightning==2.1.3

# Feature engineering and data
feast==0.35.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlmodel==0.0.14

# API and serving
fastapi==0.109.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
httpx==0.26.0

# Data streaming
kafka-python==2.0.2
confluent-kafka==2.3.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Object storage
minio==7.2.3

# ML tracking and experiment management
mlflow==2.9.2

# Explainability
shap==0.44.0
lime==0.2.0.1

# Drift detection
alibi-detect==0.11.4

# Monitoring and metrics
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Differential privacy
opacus==1.4.1

# Model calibration
scikit-learn-extra==0.3.0
netcal==1.3.5

# Utilities
python-dotenv==1.0.0
python-json-logger==2.0.7
pyyaml==6.0.1
click==8.1.7
tqdm==4.66.1
colorlog==6.8.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-xdist==3.5.0
httpx==0.26.0  # For testing FastAPI

# Code quality
ruff==0.1.11
black==23.12.1
mypy==1.8.0
pre-commit==3.6.0

# Type stubs
types-PyYAML==6.0.12.12
types-requests==2.31.0.20240106
types-python-dateutil==2.8.19.20240106

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
mkdocs-material-extensions==1.3.1
pymdown-extensions==10.7

# File watching for autopush
watchdog==3.0.0
gitpython==3.1.41

# Security
cryptography==42.0.0

# Datetime utilities
python-dateutil==2.8.2
pytz==2023.3

# Data validation
great-expectations==0.18.8

# Plotting (for explanations and monitoring)
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0

# Math and statistics
scipy==1.11.4
statsmodels==0.14.1

# Job scheduling
apscheduler==3.10.4

# Configuration management
hydra-core==1.3.2
omegaconf==2.3.0

# Progress bars
rich==13.7.0

# Jupyter (for exploration)
jupyter==1.0.0
ipykernel==6.28.0
ipywidgets==8.1.1

# Memory profiling
memory-profiler==0.61.0

